requirements.txt
# Core dependencies
click>=8.1.0
pyyaml>=6.0
jinja2>=3.1.0
pydantic>=2.0.0

# LLM and AI
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0

# Kubernetes
kubernetes>=28.1.0

# Docker
docker>=6.1.0

# SSH and remote execution
paramiko>=3.3.0
ansible>=8.0.0

# Monitoring and metrics
prometheus-client>=0.17.0
requests>=2.31.0

# Slack integration
slack-sdk>=3.21.0
flask>=2.3.0

# PDF generation
md-to-pdf-ng>=0.1.0

# Logging and utilities
structlog>=23.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 